Fixed compiler warnings
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@1160 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_callfunc.cpp
@@ -48,7 +48,7 @@ int DetectCallingConvention(bool isMethod, const asSFuncPtr &ptr, int callConv, 
 {
 	memset(internal, 0, sizeof(asSSystemFunctionInterface));
 
-	internal->func = (size_t)ptr.ptr.f.func;
+	internal->func = ptr.ptr.f.func;
 
 	// Was a compatible calling convention specified?
 	if( internal->func )
